Move notify package to cli/version
