+++ pkg/odo/cli/version/notify_test.go
@@ -1,4 +1,4 @@
-package notify
+package version
 
 /*
 import (
